pr에 대한 수정
