+++ product/src/main/java/com/everycommerce/product/domain/Product.java
@@ -5,10 +5,11 @@ import lombok.Setter;
 import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
+import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "product")
-@DynamicUpdate
+@DynamicUpdate /*모든 컬럼중에 변경된 컬럼만 update하는 기능*/
 @Getter
 @Setter
 public class Product {
@@ -44,7 +45,8 @@ public class Product {
 	/**
 	 * 등록일자
 	 */
-
+	@Column
+	private LocalDateTime createdDate;
 
 	/**
 	 * 남은수량
