+++ product/src/main/java/com/everycommerce/product/exception/InsufficientQuantitiyException.java
@@ -1,0 +1,7 @@
+package com.everycommerce.product.exception;
+
+public class InsufficientQuantitiyException extends RuntimeException{
+	public InsufficientQuantitiyException(String message) {
+		super(message);
+	}
+}
